drm/amd/display: Enable ogam_ram for dcn32+dcn35+dcn351
The comment here states "no OGAM in DPP since DCN1", yet that is not true.

Testing on an RX 7900XTX (dcn32), it seems like it actually does exist in hardware and works fine.
My best guess is the comment is confused with OGAM ROM for DPP, rather than OGAM RAM.

I did not test dcn35/351 as I do not have that hardware, but I assume the same follows there given the erroneous comment.
Someone at AMD should check that before merging this commit.

Signed-off-by: Joshua Ashton <joshua@froggi.es>